Pair iterator variable with the required type
* Editor bugged out so I didn't notice warnings about mismatches

+++ pkg/acs/lib/trim2d.c
@@ -58,7 +58,7 @@ SingleGroup *b    o: output data
 
     /* i,j = pixel indices in input array, m,n = pixel idices in output array */
     for (int n = 0, j = ystart; n < ny; n++, j++) {
-        for (size_t m = 0, i = xstart;  m < nx;  m++, i++) {
+        for (int m = 0, i = xstart;  m < nx;  m++, i++) {
             /* Extract the science array */
             Pix(b->sci.data, m, n) = Pix(a->sci.data, i, j);
